chore: refactor withRouter HoC
